# docs/swagger.json
{
  "swagger": "2.0",
  "basePath": "{{.BasePath}}",
  "host": "{{.Host}}",
  "info": {
    "contact": {},
    "description": "{{.Description}}",
    "license": {},
    "title": "{{.Title}}",
    "version": "{{.Version}}"
  },
...
  "paths": {
...
    "/v1/circles": {
			"get": {
				"tags": [
					"Circles"
				],
				"description": "Get circles",
				"summary": "Get circles",
				"produces": [
					"application/json"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "format": "byte"
          }
				],
				"responses": {
					"200": {
						"description": "Success",
            "schema": {
              "$ref": "#/definitions/CircleResponse"
            }
					},
					"400": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					},
					"422": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					},
					"500": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				}
			},
      "patch": {
        "tags": [
          "Circles"
        ],
        "description": "Update circles",
        "summary": "Update circles",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "format": "byte"
          },
          {
            "in": "body",
            "name": "circle",
            "description": "The circle to update.",
            "schema": {
              "type": "object"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Client error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "422": {
            "description": "Client error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "500": {
            "description": "Client error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
      "/v1/circles/create_circle": {
        "post": {
          "tags": [
            "Circles"
          ],
          "description": "Create a circle",
          "summary": "Create a circle",
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "circle",
              "description": "The circle to create.",
              "schema": {
                "type": "object",
                "required": [
                  "account_id",
                  "circle_symbol"
                ],
                "properties": {
                  "account_id": {
                    "type": "string"
                  },
                  "circle_symbol": {
                    "type": "string"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "schema": {
                "$ref": "#/definitions/CircleResponse"
              }
            },
            "400": {
              "description": "Client error",
              "schema": {
                "$ref": "#/definitions/ErrorResponse"
              }
            },
            "422": {
              "description": "Client error",
              "schema": {
                "$ref": "#/definitions/ErrorResponse"
              }
            },
            "500": {
              "description": "Client error",
              "schema": {
                "$ref": "#/definitions/ErrorResponse"
              }
            }
          }
        }
      },
		"/v1/circles/{id}": {
			"get": {
				"tags": [
					"Circles"
				],
				"description": "Get circles",
				"summary": "Get circles",
				"produces": [
					"application/json"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "format": "byte"
          }
				],
				"responses": {
					"200": {
						"description": "Success",
            "schema": {
              "$ref": "#/definitions/CircleResponse"
            }
					},
					"400": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					},
					"422": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					},
					"500": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				}
			},

			"delete": {
				"tags": [
					"Circles"
				],
				"description": "Delete Circles",
				"summary": "Delete Circles",
				"produces": [
					"application/json"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "format": "byte"
          }
				],
				"responses": {
					"204": {
						"description": "Success"
					},
					"400": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					},
					"422": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					},
					"500": {
						"description": "Client error",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				}
			}
		},
...
  },
  "definitions": {
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "errors": {
          "type": "object",
          "items": {
            "type": "object"
          }
        }
      }
    },
...
    "CircleResponse": {
      "type": "object",
      "properties": {
        "created_at": { 
          "type": "string"
        },
        "updated_at": { 
          "type": "string"
        },
        "deleted_at": { 
          "type": "string"
        },
        "id": 4,
        "account_id": { 
          "type": "string"
        },
        "circle_symbol": { 
          "type": "string"
        },
        "circle_name": { 
          "type": "string"
        },
        "circle_bio": { 
          "type": "string"
        }
      }
    },
...
  }
}